Move addition to separate modules
